backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
scikit-learn==1.5.1
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2; python_version < "3.13"
rapidfuzz==3.9.4
spacy==3.7.4; python_version < "3.13"
torch==2.3.1; python_version < "3.13"
transformers==4.35.0; python_version < "3.13"
tokenizers==0.14.1; python_version < "3.13"
faster-whisper==1.0.0
piper-tts==1.3.0
